initial structure
feat: functional backend with DB
