+++ backend/src/main/java/com/insurancepriceestimator/backend/repository/QuoteRepository.java
@@ -1,6 +1,6 @@
-package com.insurancePriceEstimator.backend.repository;
+package com.insurancepriceestimator.backend.repository;
 
-import com.insurancePriceEstimator.backend.entity.Quote;
+import com.insurancepriceestimator.backend.entity.Quote;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
