+++ backend/src/main/resources/application.properties
@@ -1,7 +1,7 @@
 spring.application.name=backend
 spring.datasource.url=jdbc:postgresql://localhost:5432/insurance
-spring.datasource.username=your_username
-spring.datasource.password=your_password
+spring.datasource.username=postgres
+spring.datasource.password=postgres
 spring.datasource.driver-class-name=org.postgresql.Driver
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
